homework 2
uppose you want to estimate the Earth’s proportion of water very precisely. Specifically, you want the 99% percentile interval of the posterior distribution of p to be only 0.05 wide. This means the distance be- tween the upper and lower bound of the interval should be 0.05. How many times will you have to toss the globe to do this?